Use a Bloom filter as the set for duplicate names
While the previous hash set approach is a good solution to check if an
element is in a set, the size of the data structure becomes infeasibly
large on conventional machines as the input FastQ grows, e.g., > 500
million blocks. This patch trades time for a more space-efficient
probabilistic data structure, which inherently requires two passes
through the data to guarantee correctness.