docker: Switch from musl to glibc base
This switches the image base from `alpine:3.11` (musl) to
`debian:buster-slim` (glibc). Doing simple wall time tests for ~6
million reads on `lint`, the glibc build is ~1.2x faster.

  * musl (alpine:3.11): 48.2±0.087 s
  * glibc (debian:buster-slim): 39.6±0.260 s

The difference seems to be largely in each library's implementation of
`malloc`. This can be revisited in the future when/if musl reworks its
[`malloc` implementation](https://github.com/richfelker/mallocng-draft).